Fixed games list refresh on navigation
Fixed an issue where the GamesNav component was not refreshing. This was related to dyanmic URL params in Router. I modified the Router to pass the console id property directly, and query the data based on that value.

I would like at some point to get the dyanmic URL param working, but I could not figure that out so I split up each Route path into their respective path. Will be okay for now, but not very extensible.

Backend will now query based on URL parameter.