Implement Edit and Delete Game
Edit opens a modal popup where the title can be changed. Console cannot be changed. Need to add a text section so the user can update the text.

Added a Delete button for each list item when mouse over. Clicking the delete button opens a modal dialog to confirm delete.

Changed ListView to a Component in order to manage state for current item selected. Possibly need to move this to the parent Content component or higher.