Refactored Console and Game Navigation
In an effort to reduce complexity and make the migration to React Native easier, I refactored how navigating between console and content works.

Most of the responsbility was moved to the Games component, managing the gameList and console selected.

Removed the console nav menu and replaced it with a button dropdown.

Removed the price pieces for now as well.