Clean up and move files around
Reorganized some of the file structure to split out the screens versus components in the app. Rewired App.js, MainApp.js and index.js so expo can load in web or mobile.

Added script that starts just the flask server. Hardcoded the ip of the flask server to be my desktop, expo was throwing an error when sending request to relative URL.

Games screen loads console list.