Game List Refresh
Added automatic refresh when a game is created, or the game list is scrolled up to trigger a refresh.

Tweaked some of the error handling around async request. Would like to move that function to utils, however I couldn't get the Promise response to work correctly, the error code wasn't being recognized.

Added an exception that can be thrown from the views, useful for error testing and I might need it at some point, so leaving it in for now.