Game Details Screen and Edit Game Form
Finished the layout for Game Details Screen, and passing the title to the screen through route navigate params.

Added script to open Android emulator.

Modified the game model to include a reference to console. This was to address querying a game and referencing the Console Name, since only consoleId was on the Game object.

Edit Game Form opens correctly, does not update object yet.