Improved Game List Formatting and Option Menu. Add GameImage table
Shrunk the size of each line item in game list, and added a 3 button icon for addition menu options. Currently navigates to the edit screen, but eventually it will be a dropdown/picked/select type where edit would be an option.

Added GameImage table to store pictures and screenshots for each game. Having trouble getting them to render dynamically, so checking in changes for now.